moved few things from main into separate files
